pyproject.toml
[project]
name = "copilot-memory-kb"
version = "0.2.0"
description = "Personal knowledge base compiled from Copilot CLI conversations - inspired by Karpathy's LLM KB architecture"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "tzdata>=2024.1",
    "azure-identity>=1.16.0",
]

[tool.ruff]
line-length = 100
